Move codegen module into inline_assistant
